Fix e2e test sequence verification

+++ test/e2e_test.go
@@ -279,7 +279,7 @@ var _ = Describe("e2e", func() {
 			committedChaincodes := result.GetChaincodeDefinitions()
 			Expect(committedChaincodes).To(HaveLen(1), "number of committed chaincodes")
 			Expect(committedChaincodes[0].GetName()).To(Equal("basic"), "committed chaincode name")
-			Expect(committedChaincodes[0].GetSequence()).To(Equal(1), "committed chaincode sequence")
+			Expect(committedChaincodes[0].GetSequence()).To(Equal(int64(1)), "committed chaincode sequence")
 
 			f, _ := os.Create("PackageID")
 			_, err = io.WriteString(f, packageID)
